feat(11107): Moved error consumer out of index
